requirements.txt
Flask
python-dotenv
requests
twilio
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
APScheduler
openai 